feat: main.ts에 모여있던 컴포넌트들 파일로 분리
- 각 컴포넌트들을 파일로 분리해서 만들었다.
- 전역에 선언해두고 사용했던 state는 일단 각 클래스의 인스턴스를 만들때
상위 요소가 하위요소에 props 인자로 전달하는 방식으로 했다.
- props의 type에 오류가 나고 있어서 각 타입들을 선언해야 한다.
